requirements.txt
crewai[tools]==1.6.1
python-dotenv==1.2.1
google-api-python-client==2.187.0
openai==2.8.1
lancedb==0.25.3
pandas==2.3.3
tabulate==0.9.0
openpyxl==3.1.5
gradio>=6.0.0
huggingface-hub>=1.0.0
yt-dlp==2025.11.12
pyyaml==6.0.3
pymysql==1.1.0
